requirements-dev.txt
docker-compose
django-debug-toolbar
httpie